build: centralize custom element defs and types
