--- src/elements/collator/index.ts
@@ -1,19 +1,0 @@
-import HTMLIntlCollatorElement from './collator.js';
-import HTMLIntlCollatorCompareElement from './collator-compare.js';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'intl-collator': HTMLIntlCollatorElement;
-    'intl-collator-compare': HTMLIntlCollatorCompareElement;
-  }
-}
-
-export function defineIntlCollatorElements() {
-  customElements.define('intl-collator', HTMLIntlCollatorElement);
-  customElements.define('intl-collator-compare', HTMLIntlCollatorCompareElement);
-}
-
-export {
-  HTMLIntlCollatorElement,
-  HTMLIntlCollatorCompareElement,
-}
