--- src/elements/locale/index.ts
@@ -1,15 +1,0 @@
-import HTMLIntlLocaleElement from './locale.js';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'intl-locale': HTMLIntlLocaleElement;
-  }
-}
-
-export function defineIntlLocaleElements() {
-  customElements.define('intl-locale', HTMLIntlLocaleElement);
-}
-
-export {
-  HTMLIntlLocaleElement,
-}
