+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src/"],
+  "include": ["src"],
   "compilerOptions": {
     "target": "es2019",
     "module": "es2015",
@@ -16,10 +16,6 @@
       "ES2021.Intl",
       "ES2022.Intl",
       "ESNext.Intl"
-    ],
-    "typeRoots": [
-      "node_modules/@types",
-      "./src/types.d.ts"
     ],
     "noImplicitAny": true,
     "strictNullChecks": true,
